@@ -1,6 +1,6 @@
 {
   "name": "inquirer",
-  "version": "8.1.0",
+  "version": "8.2.5",
   "description": "A collection of common interactive command line user interfaces.",
   "author": "Simon Boudrias <admin@simonboudrias.com>",
   "files": [
@@ -17,16 +17,18 @@
     "menu"
   ],
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=12.0.0"
   },
   "devDependencies": {
-    "chai": "^4.3.4",
-    "chalk-pipe": "^4.0.0",
+    "chai": "^4.3.6",
+    "chai-string": "^1.5.0",
+    "chalk-pipe": "^5.1.1",
     "cmdify": "^0.0.4",
-    "mocha": "^8.4.0",
+    "mocha": "^9.2.2",
     "mockery": "^2.1.0",
     "nyc": "^15.0.0",
-    "sinon": "^10.0.0"
+    "sinon": "^13.0.2",
+    "terminal-link": "^2.1.1"
   },
   "scripts": {
     "test": "nyc mocha test/**/* -r ./test/before",
@@ -45,12 +47,13 @@
     "figures": "^3.0.0",
     "lodash": "^4.17.21",
     "mute-stream": "0.0.8",
-    "ora": "^5.3.0",
+    "ora": "^5.4.1",
     "run-async": "^2.4.0",
-    "rxjs": "^6.6.6",
+    "rxjs": "^7.5.5",
     "string-width": "^4.1.0",
     "strip-ansi": "^6.0.0",
-    "through": "^2.3.6"
+    "through": "^2.3.6",
+    "wrap-ansi": "^7.0.0"
   },
-  "gitHead": "c4a3c2301a77b2b72c207e9d75b1061a35a1d88d"
+  "gitHead": "7a2ade6cf6a3d987f4138c0426493460f6b2515f"
 }
