@@ -91,92 +150,96 @@
   "bugs": {
     "url": "https://github.com/ReactiveX/RxJS/issues"
   },
-  "homepage": "https://github.com/ReactiveX/RxJS",
+  "homepage": "https://rxjs.dev",
   "dependencies": {
-    "tslib": "^1.9.0"
+    "tslib": "^2.1.0"
   },
   "devDependencies": {
     "@angular-devkit/build-optimizer": "0.4.6",
-    "@angular-devkit/schematics": "^0.5.4",
-    "@types/chai": "4.1.2",
+    "@angular-devkit/schematics": "^11.0.7",
+    "@swc/core": "^1.2.128",
+    "@swc/helpers": "^0.3.2",
+    "@types/chai": "^4.2.11",
     "@types/lodash": "4.14.102",
-    "@types/mocha": "2.2.48",
-    "@types/node": "9.4.5",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^14.14.6",
+    "@types/shelljs": "^0.8.8",
     "@types/sinon": "4.1.3",
     "@types/sinon-chai": "2.7.29",
     "@types/source-map": "^0.5.2",
+    "@typescript-eslint/eslint-plugin": "^4.29.1",
+    "@typescript-eslint/parser": "^4.29.1",
     "babel-polyfill": "6.26.0",
-    "benchmark": "2.1.0",
-    "benchpress": "2.0.0-beta.1",
-    "chai": "4.1.2",
-    "check-side-effects": "0.0.20",
+    "chai": "^4.2.0",
+    "check-side-effects": "0.0.23",
     "color": "3.0.0",
     "colors": "1.1.2",
-    "commitizen": "2.9.6",
-    "coveralls": "3.0.0",
     "cross-env": "5.1.3",
     "cz-conventional-changelog": "1.2.0",
-    "danger": "1.1.0",
-    "dependency-cruiser": "2.13.0",
-    "doctoc": "1.3.0",
-    "dtslint": "0.6.1",
+    "dependency-cruiser": "^9.12.0",
     "escape-string-regexp": "1.0.5",
-    "esdoc": "0.4.7",
-    "eslint": "4.17.0",
+    "eslint": "^7.8.1",
     "eslint-plugin-jasmine": "^2.10.1",
-    "fs-extra": "5.0.0",
-    "get-folder-size": "1.0.1",
+    "form-data": "^3.0.0",
+    "fs-extra": "^8.1.0",
     "glob": "7.1.2",
-    "gm": "1.23.1",
     "google-closure-compiler-js": "20170218.0.0",
-    "gzip-size": "4.1.0",
-    "http-server": "0.11.1",
-    "husky": "0.14.3",
+    "husky": "^4.2.5",
     "klaw-sync": "3.0.2",
-    "lint-staged": "7.1.1",
-    "lodash": "4.17.5",
-    "markdown-doctest": "0.9.1",
-    "minimist": "1.2.0",
-    "mkdirp": "^1.0.4",
-    "mocha": "5.0.0",
-    "mocha-in-sauce": "0.0.1",
+    "lint-staged": "^10.2.11",
+    "lodash": "^4.17.15",
+    "minimist": "^1.2.5",
+    "mocha": "^8.1.3",
+    "nodemon": "^1.9.2",
     "npm-run-all": "4.1.2",
-    "nyc": "11.4.1",
     "opn-cli": "3.1.0",
     "platform": "1.3.5",
+    "prettier": "^2.5.1",
     "promise": "8.0.1",
-    "protractor": "3.1.1",
     "rollup": "0.66.6",
     "rollup-plugin-alias": "1.4.0",
     "rollup-plugin-inject": "2.0.0",
     "rollup-plugin-node-resolve": "2.0.0",
-    "rx": "latest",
-    "rxjs": "^5.5.7",
+    "shelljs": "^0.8.4",
     "shx": "^0.3.2",
     "sinon": "4.3.0",
     "sinon-chai": "2.14.0",
     "source-map-support": "0.5.3",
-    "symbol-observable": "1.0.1",
     "systemjs": "^0.21.0",
-    "ts-api-guardian": "^0.5.0",
-    "ts-node": "6.1.0",
-    "tsconfig-paths": "3.2.0",
-    "tslint": "5.9.1",
-    "tslint-etc": "1.2.6",
+    "ts-node": "^9.1.1",
+    "tslint": "^5.20.1",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-etc": "1.13.10",
     "tslint-no-toplevel-property-access": "0.0.2",
     "tslint-no-unused-expression-chai": "0.0.3",
-    "typescript": "^3.0.1",
+    "typescript": "~4.2.0",
     "validate-commit-msg": "2.14.0",
-    "webpack": "1.13.1",
-    "xmlhttprequest": "1.8.0"
+    "web-streams-polyfill": "^3.0.2",
+    "webpack": "^4.31.0"
   },
-  "engines": {
-    "npm": ">=2.0.0"
-  },
-  "typings": "./index.d.ts",
-  "ng-update": {
-    "migrations": "./migrations/collection.json"
-  },
-  "module": "./_esm5/index.js",
-  "es2015": "./_esm2015/index.js"
+  "files": [
+    "dist/bundles",
+    "dist/cjs/**/!(*.tsbuildinfo)",
+    "dist/esm/**/!(*.tsbuildinfo)",
+    "dist/esm5/**/!(*.tsbuildinfo)",
+    "dist/types/**/!(*.tsbuildinfo)",
+    "ajax",
+    "fetch",
+    "operators",
+    "testing",
+    "webSocket",
+    "src",
+    "CHANGELOG.md",
+    "CODE_OF_CONDUCT.md",
+    "LICENSE.txt",
+    "package.json",
+    "README.md",
+    "tsconfig.json"
+  ],
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged",
+      "commit-msg": "validate-commit-msg"
+    }
+  }
 }
